# api/openapi.yaml
openapi: 3.0.3
info:
  title: ORAG API
  version: 0.1.0
  description: Go RAG framework API backed by Hertz, Eino Graph, Ark/Doubao, Qdrant, and PostgreSQL.
servers:
  - url: http://localhost:8080
...
paths:
...
  /v1/auth/login:
    post:
      tags: [Auth]
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Bearer token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /v1/knowledge-bases:
    get:
      tags: [KnowledgeBases]
      operationId: listKnowledgeBases
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Knowledge base list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeBaseListResponse"
        "401":
          $ref: "#/components/responses/Error"
    post:
      tags: [KnowledgeBases]
      operationId: createKnowledgeBase
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKnowledgeBaseRequest"
      responses:
        "201":
          description: Created knowledge base.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeBase"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /v1/knowledge-bases/{id}:
    get:
      tags: [KnowledgeBases]
      operationId: getKnowledgeBase
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Knowledge base.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeBase"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      tags: [KnowledgeBases]
      operationId: deleteKnowledgeBase
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "204":
          description: Delete request accepted by the current handler.
        "401":
          $ref: "#/components/responses/Error"
  /v1/knowledge-bases/{id}/documents:
    post:
      tags: [Ingestion]
      operationId: uploadDocument
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "202":
          description: Document ingestion accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestionResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "413":
          $ref: "#/components/responses/Error"
  "/v1/knowledge-bases/{id}/documents:import":
    post:
      tags: [Ingestion]
      operationId: importDocument
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportDocumentRequest"
      responses:
        "202":
          description: Document ingestion accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestionResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "413":
          $ref: "#/components/responses/Error"
  /v1/ingestion-jobs/{id}:
    get:
      tags: [Ingestion]
      operationId: getIngestionJob
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Ingestion job.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestionJob"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/query:
    post:
      tags: [Query]
      operationId: query
      description: |
        Executes a RAG query. The response includes `trace_id`, which is the
        same request trace propagated through logs and persisted PostgreSQL
        RAG trace records. Use `GET /v1/traces/{trace_id}` or
        `oragctl trace --trace-id <trace_id>` to inspect persisted trace
        metadata and node spans.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryRequest"
      responses:
        "200":
          description: RAG query response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  "/v1/query:stream":
    post:
      tags: [Query]
      operationId: queryStream
      description: |
        Server-Sent Events stream. Events are emitted as `trace`, `chunk`,
        `citations`, `done`, or `error`. The first `trace` event carries the
        request `trace_id`; if the RAG phase fails, the `error` event uses the
        same `trace_id` as JSON error responses and service logs.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryRequest"
      responses:
        "200":
          description: SSE query stream.
          content:
            text/event-stream:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "500":
          description: SSE error event.
          content:
            text/event-stream:
              schema:
                type: string
  /v1/traces:
    get:
      tags: [Trace]
      operationId: listTraces
      description: Lists persisted RAG traces for the current tenant with controlled filters.
      security:
        - bearerAuth: []
      parameters:
        - name: profile
          in: query
          schema:
            $ref: "#/components/schemas/Profile"
        - name: since
          in: query
          description: Inclusive lower bound for trace creation time.
          schema:
            type: string
            format: date-time
        - name: until
          in: query
          description: Inclusive upper bound for trace creation time.
          schema:
            type: string
            format: date-time
        - name: has_error
          in: query
          schema:
            type: boolean
        - name: slow_ms
          in: query
          description: Minimum trace latency in milliseconds.
          schema:
            type: integer
            format: int64
            minimum: 0
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
      responses:
        "200":
          description: Trace list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TraceListResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /v1/traces/{trace_id}:
    get:
      tags: [Trace]
      operationId: getTrace
      description: Returns a persisted RAG trace and its node spans for the current tenant.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TraceID"
      responses:
        "200":
          description: Trace record.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TraceRecord"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /v1/datasets:
    post:
      tags: [Evaluation]
      operationId: createDataset
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDatasetRequest"
      responses:
        "201":
          description: Dataset.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dataset"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /v1/datasets/{id}/items:
    post:
      tags: [Evaluation]
      operationId: addDatasetItem
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDatasetItemRequest"
      responses:
        "201":
          description: Dataset item.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DatasetItem"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /v1/evaluations:
    post:
      tags: [Evaluation]
      operationId: runEvaluation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunEvaluationRequest"
      responses:
        "202":
          description: Evaluation run result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunEvaluationResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /v1/evaluations/{id}:
    get:
      tags: [Evaluation]
      operationId: getEvaluation
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Evaluation run result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunEvaluationResponse"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/optimizations:
    post:
      tags: [Evaluation]
      operationId: optimize
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OptimizeRequest"
      responses:
        "202":
          description: Optimization result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OptimizeResult"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: |
        Unified error response. The `trace_id` is always present, is echoed from
        the incoming `X-Trace-ID` header when supplied, and is also returned in
        the `X-Trace-ID` response header for log and persisted trace correlation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: string
    TraceID:
      name: trace_id
      in: path
      required: true
      schema:
        type: string
  schemas:
    Profile:
      type: string
      enum: [realtime, high_precision]
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      required: [access_token, token_type, expires_in]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
...
    CreateKnowledgeBaseRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        metadata:
          type: object
          additionalProperties:
            type: string
    KnowledgeBase:
      type: object
      required: [id, tenant_id, name, description, created_at, updated_at]
      properties:
        id:
          type: string
        tenant_id:
          type: string
        name:
          type: string
        description:
          type: string
        metadata:
          type: object
          additionalProperties:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    KnowledgeBaseListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/KnowledgeBase"
    ImportDocumentRequest:
      type: object
      required: [content]
      properties:
        source_uri:
          type: string
        name:
          type: string
        content:
          type: string
    IngestionResponse:
      type: object
      required: [document, chunks, job]
      properties:
        document:
          $ref: "#/components/schemas/Document"
        chunks:
          type: integer
        job:
          $ref: "#/components/schemas/IngestionJob"
    IngestionJob:
      type: object
      required: [id, tenant_id, knowledge_base_id, status, source_uri, chunk_count, created_at, updated_at]
      properties:
        id:
          type: string
        tenant_id:
          type: string
        knowledge_base_id:
          type: string
        status:
          type: string
          enum: [running, succeeded, failed]
        source_uri:
          type: string
        document_id:
          type: string
        chunk_count:
          type: integer
        error:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    QueryRequest:
      type: object
      required: [knowledge_base_id, query]
      properties:
        knowledge_base_id:
          type: string
        query:
          type: string
        profile:
          $ref: "#/components/schemas/Profile"
        session_id:
          type: string
        top_k:
          type: integer
    QueryResponse:
      type: object
      required: [answer, citations, retrieved_chunks, trace_id, cache_status, profile, latency_ms, created_at]
      properties:
        answer:
          type: string
        citations:
          type: array
          items:
            $ref: "#/components/schemas/Citation"
        retrieved_chunks:
          type: array
          items:
            $ref: "#/components/schemas/SearchResult"
        trace_id:
          type: string
          description: Request trace identifier. Query traces can be inspected with `oragctl trace --trace-id <trace_id>`.
        cache_status:
          type: string
          enum: [hit, miss, error]
        profile:
          $ref: "#/components/schemas/Profile"
        warnings:
          type: array
          items:
            type: string
        latency_ms:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
    TraceListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TraceRecord"
    TraceRecord:
      type: object
      required: [trace_id, tenant_id, profile, latency_ms, created_at, has_error, error_count, node_spans]
      properties:
        trace_id:
          type: string
        tenant_id:
          type: string
        profile:
          $ref: "#/components/schemas/Profile"
        latency_ms:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
        has_error:
          type: boolean
        error_count:
          type: integer
        node_spans:
          type: array
          items:
            $ref: "#/components/schemas/TraceNodeSpan"
...
    CreateDatasetRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        kind:
          type: string
          default: golden
    Dataset:
      type: object
      required: [id, tenant_id, name, kind, version, created_at]
      properties:
        id:
          type: string
        tenant_id:
          type: string
        name:
          type: string
        kind:
          type: string
        version:
          type: string
        created_at:
          type: string
          format: date-time
    CreateDatasetItemRequest:
      type: object
      required: [query, ground_truth]
      properties:
        query:
          type: string
        ground_truth:
          type: string
        relevant_doc_ids:
          type: array
          items:
            type: string
    DatasetItem:
      type: object
      required: [id, dataset_id, query, ground_truth, relevant_doc_ids]
      properties:
        id:
          type: string
        dataset_id:
          type: string
        query:
          type: string
        ground_truth:
          type: string
        relevant_doc_ids:
          type: array
          items:
            type: string
    RunEvaluationRequest:
      type: object
      required: [dataset_id, knowledge_base_id]
      properties:
        dataset_id:
          type: string
        knowledge_base_id:
          type: string
        profile:
          $ref: "#/components/schemas/Profile"
        top_k:
          type: integer
...
    OptimizeRequest:
      type: object
      required: [dataset_id, knowledge_base_id]
      properties:
        dataset_id:
          type: string
        knowledge_base_id:
          type: string
        profiles:
          type: array
          items:
            $ref: "#/components/schemas/Profile"
        top_ks:
          type: array
          items:
            type: integer
...